Avoid running ldap backup on subclouds
LDAP should not be configured in subclouds. The bug being addressed
here caused LDAP to be erroneously be configured in controller-0 of
subclouds while not in controller-1. That caused confusion and different
testing behaviours when tests were performed against c0 and c1.

The script being fixed in this commit is one example of such
misunderstanding and confusion.

See other commits with this same topic for better comprehension.

Test cases:
PASS: In a subcloud with this change run
      '/etc/upgrade.d/11-ldap-users-backup.sh 21.12 22.12 start'.
      Verify that 'No actions required for this system type' is logged
      to /var/log/platform.log.
      Verify that 'echo $?' returns '0'.

Note: I don't think a full upgrade test is necessary for this change.
It looks straightforward to me and the testing I have performed already
ensures 'system ugprade-start' command will no longer fail because of
this script. Let me know if you disagree.

Partial-Bug: 2025663
Change-Id: I74796923b44b3ed985346940ec9a7b06da3ce336
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>