Firewall: enable IGMP for mgmt and cluster-host networks
Depending on the IGMP switch configuration between the nodes, IGMP
query and notification packets need to be exchanged. If the firewall
drops IGMP packets, the querier will not receive the IGMP report packets
and as a result multicast packets will not be forwarded by the switch.
Since hbs uses multicast, the controllers will not to able to sync
correctly.

This fix adds both ingress and egress rules to allow IGMP packets on
management and cluster-host networks.

Test Plan
[PASS] check globalnetworkpolicies IGMP rules for mgmt and cluster-host
[PASS] on each controller, check with tcpdump that it's receiving IGMP
      replies and sending IGMP reports
[PASS] on each controller, check with tcpdump that it's receiving hbs
       multicast packets from both controllers
[PASS] check with kubectl that controllers are in sync

Story: 2010591
Task: 48271

Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>
Change-Id: Idaca99a1cf774854fd340cce7f52758f053503e6