Migrate the OAM firewall to the new implementation
As part of this story, the OAM firewall will be implemented using the
sysinv.puppet.platform_firewall module, as it is done for the other
platform networks. Currently, this firewall is built using a puppet
class.

All the OAM rules are implemented in this module as they were used in
puppet, from a functional point of view there will be no difference.

During tests, it was observed that the bring-up manifest did not
update the firewall config as it requires kubectl to be available at
the moment of execution. Now the file presence of
.platform_firewall_config_required in /etc/platform/ will trigger
sysinv to apply the configuration, guaranteeing the new config apply.
This file will be generated in the stx-puppet repo (see Depends-On
link below). For the stand-by controller the agent's periodic audit
task will check for the flag file and request a firewall update.

Test Plan:
[PASS] Modify configured firewalls with kubectl edit and execute
        lock/unlock, after bring-up, check they were reset to the
        original config
[PASS] Erase a HostEndpoint and execute lock/unlock, bring-up check
        the HostEndpoint was reset to the original config.
[PASS] Erase a GlobalNetworkPolicy and execute lock/unlock, on
        bring-up check the GlobalNetworkPolicy was reset to the
        original config.
[PASS] Modify http port and check is changed in the firewall
[PASS] enable HTTPS and check is changed in the firewall
[PASS] Install without storage config and check port is not on
        the firewall
[PASS] Add storage config after installation and check port is on
        the firewall after lock/unlock
[PASS] Verify all ports open in the firewall are accessible from
        outside
[PASS] Install a System Controller and verify open ports in the
       firewall
[PASS] Execute a DOR (dead office recovery) test
[PASS] Modify the OAM IP address and verify that it reflects in
       the firewall (AIO-SX)

Story: 2010591
Task: 48277

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/886108
Change-Id: Ic226f15cb6e9376c20ba5001ae717d52f4df9474
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>