Populate DNSMasq with host entries
Host DNS records must use the ".internal" domain
to configure FQDN entries with the format of <host>.internal.

Given the complexities of supporting the multi-host reconfiguration,
this task will focus on support for a simplex system only.

This commit focuses on populating the dnsmasq.addn_hosts
with the appropriate management config for AIO-SX.

Work Items:
- addition of management config to dnsmasq.addn_hosts

Testing:
- build successful
- succesfully bring up AIO-SX and AIO-DX
- ensure changes only apply on AIO-SX
- ping system host's new aliases (with and without ".internal")
  -> also with manual removal of mgmt entries from /etc/hosts
- ensure all mgmt interfaces are up
- ping registry.local

Story: 2010722
Task: 48294

Change-Id: Iacc233a61019695bce97be36cb3b9d6a0f73cd6e
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>