Create SYSTEM_CONFIG in sysinv.common.platform_firewall
As part of story 2010591 the hard-coded values for L4 ports in
puppet were moved to sysinv.common.constants and exported there to
"system.yaml".

But the bootstrap runs prior to this file availability, hence the
port definitions are not available, and barbican failed during the
subcloud bootstrap.

To avoid that we are exporting all L4 ports that were exported to
system.yaml to the runtime.yaml file by accessing SYSTEM_CONFIG
during bootstrap.

Test Plan:
[PASS] AIO-DX standalone install/unlock/enable
[PASS] AIO-DX system controller install/unlock/enable
[PASS] AIO-SX subcloud install/unlock/enable

Closes-Bug: 2025361


Change-Id: Ieca1d6dd764622ec575fc29b73cc167cae968a5e
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>