Defining _request_timeout value
This commit set an timeout of 30s to read_namespaced_secret.
Setting it to 30s allow the system to retry at least three times,
what would set the total time of the operation to 3x the timeout
set.

Test Plan:

PASS: Build an iso, and boot a system with the changes.
PASS: Benchmark test to set the more reasonable value for the timeout.

Closes-Bug: 2026759

Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>
Change-Id: I1295e93e816faff30a02084acafbb2d9ac6d944d