Remove unused container images as last step in K8s upgrade complete
In order to reduce the number of container images that we're storing,
and to reduce the amount of data that needs to be preserved for both
backup/restore and for platform upgrades, we want to add a step at the
end of the "system kube-upgrade-complete" command to delete container
images that are associated with older versions of K8s and are no
longer needed or used by target version.

This includes two categories:
1. Images specified by '/usr/local/kubernetes/<version>/stage1/usr \
   /bin/kubeadm config images list --kubernetes-version <version>'.
2. Images specified in Ansible playbooks under playbooks/roles \
   /common/load-images-information/vars/<version>/system-images.yml

Test Plan:
PASS: Perform platform upgrade and then k8s upgrade.
PASS: Perform k8s upgrade, after the step system kube-upgrade-complete
      verified that the images were deleted by randomly selecting the
      image and passing it to the command
      system registry-image-delete <image:tag> and confirming that
      output of system registry-image-tags <image_name> for respective
      image is empty
PASS: Package builds successfully

Story: 2010565
Task: 48366

Change-Id: Ifcf37fd27bb273c9d5f7bb09e6a2805b43597fe5
Signed-off-by: Saba Touheed Mujawar <sabatouheed.mujawar@windriver.com>