Add missing organization to certificate's subject
In the sysinv conductor manager, the software that builds
and updates the kube-apiserver-kubelet-client certificate
fails to add the correct kubernetes organization to the
subject.  As a result, when the user follows the procedure
to update the root ca, the new certificate is created and
not associated with the necessary subject.  Finally, kube
apiserver access is affected and the user loses the ability
to interface with kubernetes via kubecdtl commands, and the
sysinv software automation similarly.

This correction adds the appropriate subject to the
certificate.

Test Plan:
PASS: build-pkg -a && build image
PASS: AIOX-SX full install with clean bootstrap
PASS: Perform manual root ca update, and verify
      that the ability to kubectl exec into a
      pod continues to work properly.
PASS: Perform manual root ca update, next
      install an application and verify that
      the installation succeeds.

Closes-Bug: 2027714
Change-Id: Ic153f37959340e1542d85fca6d15f494192dd8a7
Signed-off-by: Joshua Reed <joshua.reed@windriver.com>