Allow ICMP(v6) to access the admin network
The current admin network firewall does not allow ICMP(v6) traffic to
the admin network itself (only the system controller), preventing the
exchanging of neighbor information with the gateway. It also
prevents network troubleshooting, by blocking ping from the gateway.

This change adds the admin network (and the link-local network for
IPv6) to the ICMP(v6) protocol, in the admin network firewall.

Closes-Bug: 2027827

Change-Id: If88ca060605b917332c35eab9daa62eb5af7085a
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>