Create TLS 'system-local-ca' in openldap_ca cert mode
Change the behavior in sysinv API when a certificate install is called
using 'openldap_ca' mode. Before, this call would recreate the secret
'system-local-ca' using the same type (Opaque or TLS) as it had
before. Now the call will recreate always as TLS, as this is required
to deployment configurations that try to create a local issuer with
'system-local-ca' secret. The call will also create a default issuer
using the secret.

This API cannot be called directly from CLI in production environment,
and it's only called as of today during a DX subcloud upgrade by
orchestration from the SystemController already in the N+1 load to the
subcloud in N load.

Test Plan:
PASS: Enable manually the CLI call for mode openldap_ca in sysinv API
      code in:
      https://opendev.org/starlingx/config/src/branch/master/sysinv/cgts-client/cgts-client/cgtsclient/v1/certificate_shell.py#L86
      Create a certificate + key bundle and install as openldap_ca.
      Check if the 'system-local-ca' secret in namespace cert-manager
      is now TLS and contains the certificate and key installed.
      Check if the 'system-local-ca' ClusterIssuer is created and
      available.

Closes-Bug: 2029508

Change-Id: Ief932778f7e2dd4e4bd703ef009ec4058b7416f0
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>