Allow firewall runtime update to be requested from any node
During patch apply tests, it was detected that if all the firewall
puppet classes are not provided during boot, we need to mark the
filesystem with .platform_firewall_config_required in /etc/platform/
so sysinv-agent can request a runtime execution to conductor. The
audit operation already looks for this file, but it was limited to
the stand-by controller.

In all test scenarios the validation was done with the verification
that the firewall was created after patch application.

Test Plan:
[PASS] Apply the firewall feature as a patch in AIO-SX
[PASS] Apply the firewall feature as a patch in AIO-DX
[PASS] Apply the firewall feature as a patch in Standard
[PASS] Execute DOR test in a Standard Install and validate firewall
       installation in worker nodes

Story: 2010591
Task: 48460

Change-Id: Ie02b9c945dd5ad55333055e9d5d36ed963be1c75
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>