Trigger stevedore cache refresh on sysinv start
This is a workaround for re-using stale stevedore cache
on sysinv-conductor startup.

The stevedore cache is immediately reloaded during the
module import stage based on a calculated digest of the
last modified timestamp of all of the paths in the python
system path, as well as all eggs it detects.

When applying a StarlingX ostree patch, the timestamps are
not updated due to how the binary contents are created,
and many folders show an older 1970 datestamp.

This means that when the conductor starts, it will think
that the old digest is still valid, and will not reload
the entry points from the updated eggs.

Most paths in the python path are immutable, and can not
be altered by a service file, but "/var/stx_app/plugins"
can be updated, and is in the python path.

By ensuring the timestamp of that folder is updated, the
 on-disk stevedore cache will be re-generated under the
/root/.cache/python-entrypoints directory.

Caveat:  every time sysinv-conductor is restarted, that
folder will create a new entry of about 200KB in size.

The reason this change needs to be done in this startup
script is because if it is inserted in the init or start
method of the conductor, the imports will have already
been invoked and the stale cache will have already been
re-loaded.  This location ensures the cache will be
re-created prior to the conductor starting.

The main area this fixes are the calls to:
stevedore.extension.ExtensionManager

which is used by the puppet loader, but also load-import.
This does not directly fix sysinv-agent lldp which also
uses that stevedore code however that list of entry points
is un-likely to change.

Closes-Bug: #2028403
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I005b3fefbe29ee426003a823f6fbb043a5059b89