Reduce pagination limit for registry requests
In the past, we introduced a large pagination limit for requests to
the registry in order to get around the default 100 results. The
default 100 results was not enough once there are more than 100
images in the registry. Recently, a change in the registry started
rejecting requests that specifies a pagination of over 1000:
https://review.opendev.org/c/starlingx/integ/+/886884

This commit changes the pagination limit from 100000
to 1000. This should still be sufficient, since it is 10x as much as
the default 100. We also do our image-list and image-tags commands
separately (busybox:1.36 and busybox:latest shows up as just busybox
on image-list. image-tags would show 1.36 and latest, but only show
tags for busybox since image-tags needs the user to specify the
image name). This means a system that have been around for a long time
and went through many upgrades would still likely not hit the limit
even if it has multiple versions of the same image.

Test Cases:

PASS: system registry-image-list

Change-Id: Ib1c071f6f66788065423f7e512366057a173cce6
Closes-Bug: 2028566
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>