Steps for kube-upgrade-storage
This adds the kube-upgrade-storage support steps on for k8s upgrade.
On the kube_upgrade_shell, the command for the upgrade-storage step
was added, and on the other files, all necessary variables and
possible states during the upgrade were added, while adding the
support for the ansible-playbook that upgrades the storage images.

With this, the expected sequence on AIO-SX is:
 - system kube-upgrade-start <target-version>
 - system kube-upgrade-download-images
 - system kube-upgrade-networking
 - system kube-upgrade-storage
 - system kube-host-cordon controller-0
 - system kube-host-upgrade controller-0 control-plane
 - system kube-host-upgrade controller-0 kubelet
 - system kube-host-uncordon controller-0
 - system kube-upgrade-complete
 - system kube-upgrade-delete

Test Plan:
PASS: Run a kubernetes upgrade with the kube-upgrade-storage step and
observe that the image for volume-snapshot-controller changed in
AIO-SX and AIO-DX configurations.
PASS: Run a kubernetes downgrade and observe that the image for
volume-snapshot-controller changed.

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/887797
Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/889393
Depends-On: https://review.opendev.org/c/starlingx/nfv/+/889390

Story: 2010877
Task: 48589

Signed-off-by: Luiz Felipe Kina <LuizFelipe.EiskeKina@windriver.com>
Change-Id: Ia04fac28505e5a0636d6453c69a4691f6c9f9e66