Sorting devlinks to avoid unnecessary updates
During sysinv-agent testing, it was noticed that sysinv reports system
changes after just reading the list of disks, sending a report as if the
system had changed, which is not the case.

Sysinv-agent logs show that pyudev is returning all the device links
(DEVLINKS) in an inconsistent order between calls. This results in the
agent thinking that disks have changed and generates unnecessary disk
update messages to the sysinv-conductor. Sorting the links returned
from pyudev, eliminates this variability.

Test Plan:
    PASS: After changing the code and restarting sysinv-agent,
          no more RPC calls to the idisk_update_by_ihost method
          and no more device_id updates to the i_idisk table in
          the sysinv database on each run.

Closes-bug: 2028802

Change-Id: I2b271b1818cf7aff189ad7025431664c864fe94b
Signed-off-by: Hediberto Cavalcante da Silva <hediberto.cavalcantedasilva@windriver.com>