Run audit only after host is inventoried
If host is not yet inventoried, the audit might run, detect
the need to reconfigure the host while the host is still unable
to receive puppet manifest apply requests, leaving a divergence
between current config and target config on the host, and
therefore a stuck alarm.

This commit allow the security audit to run only after the
host is ready to serve these requests.

Test Plan
PASS: install/bootstrap/unlock successfully with no alarms
PASS: verify the audit running after host is inventoried

Closes-bug: 2028173

Change-Id: Id75cef7ef144385e77b326eee79d5a296697db79
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>