Fix dev search on validate-platform-backup script
During testing, validate-platform-backup.sh could not find the
parent path of the nvme0n1p4 partiton, and resulted in failure
to validate.

This fix adds an additional step to extract the parent path (disk)
of the "nvme0n1p4" partition (or any similar partition format).
This is done by removing the  trailing "p" and the number at the end
(partition number) using sed. Then, we check if the parent path is
not empty and set the device_path variable to the parent path
("/dev/nvme0n1" in this case) if the condition is met.

Test Plan
PASS: Fresh Install and Bootstrap (AIO-SX)
PASS: validate-platform-backup.sh returns SUCCESS

Closes-bug: 2028943

Change-Id: I313bcca2720512fd1457b6c622bb6a2ddf2c8a68
Signed-off-by: Matheus Guilhermino <matheus.machadoguilhermino@windriver.com>