Fix error when running k8s cert rotation on c1
Review https://review.opendev.org/c/starlingx/config/+/884627
introduced parameter --config /etc/kubernetes/kubeadm.yaml for
kubeadm certs check-expiration, which generates a better output
without warnings, that are print when run without.

File /etc/kubernetes/kubeadm.yaml, however, is not present on
controller-1 so this is not a good solution. This commit removes the
usage of the parameter so that the script can work on both environments.

This script runs as a cron job only on controller nodes.

PASS: Trigger execution of script kube-cert-rotation.sh on c0, verify
      that it runs without error. Run 'fm alarm-list' and verify no
      'Kubernetes certificates renewal failed are present'
PASS: Trigger execution of script kube-cert-rotation.sh on c1, verify
      that it runs without error. Run 'fm alarm-list' and verify no
      'Kubernetes certificates renewal failed are present'

Closes-Bug: 2029506
Closes-Bug: 2029378

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I1ea458163dc4250a3bc3550eaaa68d314224023e