Update rule in vim endpoints recreation script
19-recreate-vim-systemcontroller-endpoint.sh was designed to handle
the clean-up for vim/nfv openstack endpoints for the release where
the fix for LP #1995951 was delivered.

This version was coded inside the script. Considering this, it has
to be updated to be executed for future releases. Changed it to
consider the FROM version equals 22.12 to be executed.

If other upgrade paths are to be added, this script must be updated
accordingly in the future. This script is designed to make changes
only in a DC SystemController.

PASS: After the upgrade finishes, vim/nfv should have 3 endpoints
      (public, internal and admin) in SystemController region.
      For each interface, the endpoints must be configured with
      same protocol (http or https) and base IP as the other
      services counterparts (E.g. Patching). Tested:
      - Upgrade standalone AIO-SX.
      - Upgrade DC SystemController.

Closes-bug: 1995951

Change-Id: I2cd0ef5cade189ebec153b12278494a1a834be69
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>