Better control over chart list generation.
The previous implementation of the _get_list_of_charts
method would not take into account whether or not a
particular application chart was enabled or disabled.
Instead the logic would rely on the resources present in
the top-level application kustomization.yaml manifest to
determine what FluxCD charts are enabled. As this manifest
is changed dynamically based on the the application plugin
use of the FluxCDKustomizeOperator, the resulting top-level
application kustomization.yaml could change from one
overrides generation call to the next.

This fix ensures that:

1. When all charts are needed then an option can be specified
   (i.e. when determining all the container images needed for
   the application)
3. All possible charts, as filtered by the metadata/user
   driven and DB stored enabled status, are consistently
   returned regardless of the current state of the top-level
   application kustomization.yaml.

Test Plan:
PASS: build-pkgs -a && build-image
PASS: AIO-SX full install with clean bootstrap
PASS: Enable the cms-replication chart on the dell-storage app
PASS: Use system helm-override-update to pass
      --set config.clusterID=ClusterA
PASS: system application-apply dell-storage
PASS: Check the YAML structure of the
      configmap/dell-replication-controller-config
      for ClusterA and properly formatted.

Closes-Bug: 2029303

Change-Id: Ia737322100159d342a4ac4f2471bb3f0c18adafb
Signed-off-by: Joshua Reed <joshua.reed@windriver.com>