RPC timeout when changing the system timezone
This fix increases the 'configure_system_timezone' command timeout from
60s, which is the default, to 90s. This command configures the
controllers and the workers separately, so the original RPC coming from
the Api only returns after the configuration is completed in the
controllers and after the command is sent to the workers, which can take
more than 60s in some environments that have Openstack deployed,
especially after a swact. With Openstack installed, the command
execution time increases considerably, being one of the main
contributors to this timeout.

Test Plan:
Pass: Manually changed the code in a Standard lab with Openstack
deployed, setting the timezone after a swact.
Pass: Tested in a Standard VM installation, simulating the timeout with
a sleep and checking that the change fixes the 60s threshold problem,
running the tests locally.

Closes-Bug: 2028280

Change-Id: I6d2de89b0c0ab83da02206fb1ca2b6c69d846c7b
Signed-off-by: Douglas Gameiro Diniz <douglas.diniz@windriver.com>