Add NAT rule for subclouds using the admin network
For subclouds using the admin network, the connection to the
system controller operates exclusively via the admin network.

In this configuration, only the controllers of a subcloud
have an admin network and associated interface.

The management network of the subcloud is essentially private
and is used by other nodes of the system (ie. worker) to
communicate internally.

One problem posed by this is LDAP authentication.  Because the
LDAP server is located on the system controller, a worker node
(without an admin network connection) will have an issue
connecting to the server.

The worker node _does_ have a default route to the subcloud's
controller nodes.  The controller nodes _do_ have a route to
the system controller.  The issue occurs in the opposite
direction.  Because the subcloud's management network is
essentially private, the system controller will not know how
to send responses to the worker node.

This commit in part resolves this issue by enabling an
LDAP SNAT rule which will translate worker traffic to the
LDAP server to the appropriate admin network address.  This
way, the return traffic from the system controller will know
where to go.

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/891511

Story: 2010319
Task: 48616

PASS: NAT rule is created on fresh install w/ admin network
      present.
PASS: NAT rule is deleted on deletion of admin network
      address pool.
PASS: At runtime, update the subcloud between usage of
      admin / management network.

Change-Id: I82678ebfb431e3dc46635025eeff1eb3655802d1
Signed-off-by: Steven Webster <steven.webster@windriver.com>