Retry get openstack users when unauthorized
When updating the sysinv user's password, the cached keystone client
may failed to get the openstack users due to authentication failure
(this update currently only happens during subcloud rehoming). As a
result, the kesytone listener fails to trigger the puppet manifest to
update the sysinv password on the other nodes even if the service
configuration is already updated in the active controller.

This commit detect this type of exception during getting the openstack
users, refreshes the cached keystone client and retries to get the
user list, ensures the get users method using the latest service
configuration if the unauthorized expection occurs.

Test plan:
Case 1 -- passed:
1. Rehome a large number of subclouds.
2. Verify the warning about retry by unauthorized exception.
3. Verify there's the retry was successful and the manifest was
triggerred.

Case 2 -- passed:
1. Update the sysinv password w/o updating the configuration in
advance.
2. Verify the warning about the retry by unautorized exception.
3. Verify the error about the failure in the retry.

Partial-bug: 2031182
Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I0958b6c7c7bb0ba27ef11d2bafabea40fae6b878