Restrict admin type interface to controllers nodes
Installing a DX+ system, it is possible to create an interface
and assign it to a network of type admin for non-controller nodes.

The admin interface associated to admin network should be supported
only for controller nodes. Worker nodes should reject the assignment
of a platform interface to a network of type admin.

Test Plan:
- Build successful
- Successfully bring up DC+ system
- Create interface with admin network type
- Verify worker node rejected the assigned interface

Closes-Bug: 2031024

Change-Id: I80886baa5d2d3dd45c4e67fb1b84d519924c8be3
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>