Delete kube-expired-kubelet-cert-recovery.sh script
This commit deletes kube-expired-kubelet-cert-recovery.sh.
I have refactored this script to be done by the ansible playbook.
This script is not called by anything else. It was added solely to be
used with the playbook. After hitting some issues with this
implementation, I decided it made sense to refactor it to ansible to
have more control over each individual step.

Also, this commit fixes an error on kube-cert-rotation.sh,
which runs on controller nodes only as a cron job, when
cert-rotation is performed on c1. File /etc/kubernetes/kubeadm.yaml,
is not present on c1, so the script will fail to complete with a
'file not found' error. This flag was added in review
https://review.opendev.org/c/starlingx/config/+/884627
because it returns a better output for 'kubeadm certs
check-expiration, without the many warning that are print
without. Other than the output, the script works correctly without
the --config flag.

PASS: Full build and install.
PASS: Make k8s certificates expire and then on both c0 and c1:
      - Trigger execution of script kube-cert-rotation.sh, verify
      that it runs without error. Run 'fm alarm-list' and verify no
      'Kubernetes certificates renewal failed are present' and
      that k8s certificates are now valid.

Story: 2010815
Task: 48700

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I06f01529a264eaceb0c0bb90e3947d2e0cb237d0