Improve update of user config robustness
When a user's password expiry period is updated many times per the
execution of 'chage -M [number_of_days] [user]' command, the update
user config function present in conductor/manager.py triggers the
users runtime manifest to be executed by Puppet in each one of these
command executions and this could cause a conflict in this update
process.

In this case, the sysinv conductor could call sysinv agent to apply
the manifest and queue a user config update and, before this action
is completed and marked as completed, the sysinv agent is called again
by sysinv conductor per the request of other 'chage' command execution,
therefore, provoking a deadlock in the process.

This commit improves the robustness for the update user config process
by deferring users runtime manifest executions per the check if a
manifest is currently being applied. If a users runtime manifest is
being applied and it is requested a new apply of the manifest, the
incoming request is now queued and executed after the current one is
finished.

Test Plan:
PASS: Issue the command 'sudo chage -M [number_of_days] sysadmin' many
      times, one time right after the other one and with different
      number of days. Observe in /etc/shadow if sysadmin user has the
      last addressed number of days as its password expiry period.
      Then, observe in /var/log/auth.log that the log info related to
      the update of sysadmin's password expiry was completed in every
      'chage' command issued and the deadlock was not reproduced.
      Verify in /var/log/sysinv.log that the update_user_config()
      function was successfully executed with exit status 0 for each
      time the 'chage' command was issued and users runtime manifest
      was executed successfully by Puppet as well.

Closes-Bug: 2034446

Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>
Change-Id: I769c99b870f2a8c4a6fb99e5739da1e420606ddd