Create runtime_config table
This commit creates a runtime_config table, to allow
storing runtime config entries per host so that a follow-up
commit will add the capability to retry runtime config
requests to hosts that are taking too long to report back
to conductor the runtime manifest application success.

The table added on this commit also comes with a time-based
pruning function, since it has potential to become large
depending the size and number of nodes of the deployment, so
that this function can be used by the follow-up commit to keep
only the entries inside a determined and relevant timeframe.

Test Plan:
PASS: fresh install/bootstrap/unlock single and
         multi-node deployments
PASS: DX upgrade stx-6 to stx-8
PASS: SX upgrade stx-6 to stx-8

Partial-Bug: 2037090

Change-Id: I40084fe34e8827262dc0453298626ac7394e6bb5
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>