Protection from invalid label values.
In api/controllers/v1/label.py added a guard to prevent
and null being used for a new label value or updating a
label value.  The kubeapi protects against invalid label
regex, but this check goes a step further to prevent the
possibility of a null value in the body payload of the API
post.  Previously, a null could be entered into the database.

Enhance protection of sysinv/puppet/kubernetes.py
function _get_host_label_config such that there are
checks for NoneType before string manipulation occurs.
This prevents this section of code from breaking if the
data base is missing or has a label with a null value.

Test Plan:
1. PASS: Full AIO-SX install.
2. PASS: Verify the system does not allow a null value
   to be entered when adding a new label.
3. PASS: Verify the system does not allow a null value
   to be entered when updating an existing label.
4. PASS: Delete all labels, and verify that hieradata for
   host configuration is generated properly. Verify that
   the "host_labels" puppet parameter has an empty list:
   [].

Closes-Bug: 2032572
Change-Id: I7040155616aa63986d0c37ba967a207b5c31addf
Signed-off-by: Joshua Reed <joshua.reed@windriver.com>