Add error message to system registry-image-delete
Add an error  message to registry-image-delete, currently when users
attempt to delete a non-existent image, the command fails silently.
This review is adding an error message to this operation, now when
the user tries to delete an image that doesn't exist it will be displayed
this message: "No tag found for image: <image_name>"

Also improving 'system registry-image-list' by adding an optional
argument "--filter-out-untagged" to filter and display only images
that have associated tags in the system registry.

Test Plan:
PASS: Run 'system registry-image-list' without '--filter-out-untagged'.
PASS: Run 'system registry-image-list' with '--filter-out-untagged' and
      verify if the untagged images are out of the list.
PASS: Delete an image that doesn't exist and verify if it will return
      the error message.
PASS: Delete an existing image and verify if it was successfully
      deleted.

Closes-Bug:2037552

Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>
Change-Id: Ibbc25dff5f4c1b22d6bf923235e6d0cd821b4953