Update logic for setting initial ptp4l pmc paramters
Rework the logic for how pmc GrandMaster parameters are
initialized.

The new logic checks for the presence of each of the valid Announce
message keys in the user config and updates the values for any that are
defined.

These values are then made available in the resulting hieradata and
subsequently written by puppet.

Test Plan:
PASS: Verify that the pmc GRANDMASTER_SETTINGS can be set/unset via user
supplied parameters

Partial-bug: 2036431

Change-Id: Ib4429e2f4b2477273c421a862dadb1dbede4aaac
Signed-off-by: Cole Walker <cole.walker@windriver.com>