For OAM firewall support old host-endpoint label notetype
It was observed that during upgrade that the new OAM firewall,
generated by controller-1 already in the new version, does not
contain the host endpoint label "notetype", used in the previous
versions. This label was renamed "nodetype" as it only contains the
values of 'controller' or 'worker'.

This changes adds "notetype" as a supported label in the OAM Global
Network Policy, to allow firewall application during upgrade in the
controller-0, when still running the old version. The other network
types did not had firewall attached to them in previous versions.

Test Plan
[PASS] Execute upgrade on an AIO-DX system and validate that after the
        unlock of controller-1 the OAM access is operational in
        controller-0 and controller-1
[PASS] Execute boostrap install for AIO-DX system and validate that
        that the OAM access is operational in both controllers

Closes-Bug: 2038306

Change-Id: Ib6bc95db9a2ec22ec63f693bd4f85dad89023a2b
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>