Fix uuid update in the DB partition table
The changes from [1] review caused a bug. When there is a
pv created in a partition, if the partition uuid in the
partition table  updates, the same field in i_pv is not
updated automatically, causing audit failure.

For this reason, the code related to the uuid update of a
partition in the sysinv database according to agent report was
reverted. So the DB partition uuid is the same since the register
is created and there is no chance of inconsistencies in the
sysinv database related to this.

[1]: https://review.opendev.org/c/starlingx/config/+/889090

Test Plan:
  PASS: AIO-SX fresh install.
  PASS: AIO-SX -> create/modify/delete a partition in the
        controller-0 followed by a reboot and check the status with
        'system host-disk-partition-list'.
  PASS: Check that the uuid of a partition with a pv remains the
        same in i_pv and partition tables in the sysinv database.
  PASS: Succesfull DX upgrade from stx 6.0 to stx 8.0.

Closes-Bug: 2028254

Change-Id: I08fb3610b98bed7a2531584c8f542058976894b3
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>