Check/issue platform certs in DX upgrade
Included upgrade script to verify the existence and issue if necessary
the now (after this Story) required platform certificates (REST API &
Web Server, Docker Registry and local OpenLDAP), using the
'system-local-ca' ClusterIssuer for DX systems.

These changes are dormant. The upgrade script will not be triggered
unless a specific file used as feature flag is present in the system.
This will prevent interfering with current behavior until the whole
feature is completed.

The proper system upgrades tests will be done together when the
support for DC systems is concluded in a future task for this Story.

Test plan:
PASS: Deploy AIO-SX and AIO-DX, providing the CA cert in
      'system-local-ca' overrides and the flag. Verified that:
      - HTTPS is enabled correctly after unlocking the controller.
      - The certificate under '/etc/ssl/private/' is correct.
      - HTTP is disabled correctly after deleting the certificate and
        using the 'system modify' API to disable it.

PASS: Execute the upgrade script manually and verify that the required
      platform certificates are not altered.

PASS: Delete the required platform certificates. Execute the upgrade
      script manually and verify that the required platform
      certificates are issued.

Story: 2009811
Task: 48891
Depends-on: https://review.opendev.org/c/starlingx/ansible-playbooks/+/897364

Change-Id: Ie628f24ce11fe7ad5aafb1e526320a4e943be547
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>